move model to it's own package
